% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTextGridLabels.R
\name{addTextGridLabels}
\alias{addTextGridLabels}
\title{Combine data frames with signal information and TextGrid labels}
\usage{
addTextGridLabels(signal, TextGrid, soundExt = ".wav", tgExt = ".TextGrid")
}
\arguments{
\item{signal}{A data frame with signal data, e.g. formants or pitch,
generated with the corresponding `praatutils` functions.}

\item{TextGrid}{A data frame with TextGrid data generated with
`readTextGrid()`.}

\item{soundExt}{String giving the file extension of the audio files used
to generate `signal`; default is `.wav`. This is used to match file
names across the two input data frames.}

\item{tgExt}{String giving the file extension of the TextGrid files used
to generate `TextGrid`; default is `.TextGrid`. This is used to match file
names across the two input data frames.}
}
\value{
A data frame with column names corresponding to `signal` and
`TextGrid`, and row numbers corresponding to `signal`.
}
\description{
Given a data frame with signal information and a data frame with TextGrid
labels, combine the two such that the TextGrid labels for each time step
is appended to the signal data.
}
\examples{
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
tgFile <- paste0(datapath, '/1.TextGrid')
pit <- pitchRawAC(soundFile, output = 'df')
tg <- readTextGrid(tgFile)
res <- addTextGridLabels(pit, tg)
}
