% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formant2ssff.R
\name{formant2ssff}
\alias{formant2ssff}
\title{Format Formant object as SSFF}
\usage{
formant2ssff(fm, fmTab, returnBw = FALSE, returnIntensity = FALSE)
}
\arguments{
\item{fm}{`parselmouth.Formant` object}

\item{fmTab}{Table of formant values created under the hood with
`formantBurg`.}

\item{returnBw}{Boolean; should formant bandwidths be returned? Default is
`FALSE`.}

\item{returnIntensity}{Boolean; should intensity values be returned?
Default is `FALSE`.}
}
\value{
List object of class `AsspDataObj`.
}
\description{
Convert Praat Formant object (as loaded in memory through the Python
interface `parselmouth`) to `list` object with the Simple Signal File Format.
}
\examples{
# Don't use directly
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
fm_ssff <- formantBurg(soundFile, output = 'ssff')
}
