% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formant2table.R
\name{formant2table}
\alias{formant2table}
\title{Format Formant object as table}
\usage{
formant2table(
  fm,
  filename,
  returnBw = FALSE,
  returnIntensity = FALSE,
  times = NULL
)
}
\arguments{
\item{fm}{`parselmouth.Formant` object}

\item{filename}{String giving the file name of audio file.}

\item{returnBw}{Boolean; should formant bandwidths be returned? Default is
`FALSE`.}

\item{returnIntensity}{Boolean; should intensity values be returned?
Default is `FALSE`.}

\item{times}{Optional numeric vector giving the times to return signal
values from.}
}
\value{
Data frame with formant values.
}
\description{
Convert Praat Formant object (as loaded in memory through the Python
interface `parselmouth`) to data frame.
}
\examples{
# Don't use directly
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
fm_df <- formantBurg(soundFile, output = 'df')
}
