% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonicity2df.R
\name{harmonicity2df}
\alias{harmonicity2df}
\title{Format Harmonicity object as data frame}
\usage{
harmonicity2df(hnr, filename, times = NULL)
}
\arguments{
\item{hnr}{`parselmouth.Harmonicity` object}

\item{filename}{String giving the file name of audio file.}

\item{times}{Optional numeric vector giving the times to return signal
values from.}
}
\value{
Data frame with harmonics-to-noise-ratio values.
}
\description{
Convert Praat Harmonicity object (as loaded in memory through the Python
interface `parselmouth`) to data frame.
}
\examples{
# Don't use directly
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
hnr_df <- harmonicityAC(soundFile, output = 'df')
}
