% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonicity2ssff.R
\name{harmonicity2ssff}
\alias{harmonicity2ssff}
\title{Format Harmonicity object as SSFF}
\usage{
harmonicity2ssff(hnr)
}
\arguments{
\item{hnr}{`parselmouth.Harmonicity` object}
}
\value{
List object of class `AsspDataObj`.
}
\description{
Convert Praat Harmonicity object (as loaded in memory through the Python
interface `parselmouth`) to `list` object with the Simple Signal File Format.
}
\examples{
# Don't use directly
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
hnr_ssff <- harmonicityAC(soundFile, output = 'ssff')
}
