% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonicityAC.R
\name{harmonicityAC}
\alias{harmonicityAC}
\title{Estimate harmonics-to-noise ratio using raw autocorrelation}
\usage{
harmonicityAC(
  filename,
  output = "ssff",
  ext = ".wav",
  timeStep = 0.01,
  start = NULL,
  end = NULL,
  times = NULL,
  floor = 75,
  silenceThreshold = 0.1,
  periodsPerWindow = 1,
  filterFrom = 0,
  filterTo = 0,
  filterSmoothing = 100,
  toFile = FALSE,
  outputDir = getwd(),
  outputExt = ".hnr"
)
}
\arguments{
\item{filename}{String giving the path to an audio file or a directory of
audio files. Alternatively a vector of strings giving the paths to audio
files.}

\item{output}{String giving the desired output format. Should be either
`ssff` (default) for a list with the Simple Signal File Format or `df` for a
data frame.}

\item{ext}{String giving the file extension for audio files; default is
`.wav`. Only used if `filename` is a directory.}

\item{timeStep}{Numeric giving the desired sampling frequency in
seconds. Default is `0.01`.}

\item{start}{Numeric giving the desired start time of analysis in
seconds. Default is `NULL`, equivalent to the start of the audio file.}

\item{end}{Numeric giving the desired end time of analysis in seconds.
Default is `NULL`, equivalent to the duration of the audio file.}

\item{times}{Optional numeric vector giving the times to return signal
values from. Default is `NULL`.}

\item{floor}{Numeric giving the lowest expected pitch frequency in Hz.
Default is `75`.}

\item{silenceThreshold}{Numeric giving the threshold below which frames are
considered silent. The value is a proportion of the global maximum
amplitude. Default is `0.1`. Silent frames are returned as `-200`.}

\item{periodsPerWindow}{Numeric giving the minimum number of pitch periods
to include per window. Default is `4.5`, which then sets the actual window
length with reference to `floor`.}

\item{filterFrom}{Numeric giving the lower frequency bound of an optional
bandpass filter applied to the signal before estimating HNR in Hz.
Default is `0`.}

\item{filterTo}{Numeric giving the upper frequency bound of an optional
bandpass filter applied to the signal before estimating HNR in.
Default is `0` (interpreted as "no upper bound").}

\item{filterSmoothing}{Numeric giving the width of the frequency width of
the window}

\item{toFile}{Boolean; if the output format is `ssff`, should these
objects be saved to disk as individual files? Default is `FALSE`.}

\item{outputDir}{String giving the location of the output directory of
SSFF files when `toFile = TRUE`. Default is to use the current working
directory.}

\item{outputExt}{String giving the file extension of SSFF files when
`toFile = TRUE`. Default is `.fmt`.}
}
\value{
A list of class `AsspDataObj` or data frame depending on the value
of `output`.
}
\description{
Wrapper for Praat procedure that estimates harmonicity (harmonics-to-noise
ratio) values from an audio file or list of audio files based on the raw
autocorrelation of the signal.
}
\examples{
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
hnr <- harmonicityAC(soundFile)
}
