% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity.R
\name{intensity}
\alias{intensity}
\title{Estimate root-mean-squared amplitude}
\usage{
intensity(
  filename,
  output = "ssff",
  ext = ".wav",
  timeStep = NULL,
  start = NULL,
  end = NULL,
  times = NULL,
  minPitch = 100,
  subtractMean = TRUE,
  toFile = FALSE,
  outputDir = getwd(),
  outputExt = ".rms"
)
}
\arguments{
\item{filename}{String giving the path to an audio file or a directory of
audio files. Alternatively a vector of strings giving the paths to audio
files.}

\item{output}{String giving the desired output format. Should be either
`ssff` (default) for a list with the Simple Signal File Format or `df` for a
data frame.}

\item{ext}{String giving the file extension for audio files; default is
`.wav`. Only used if `filename` is a directory.}

\item{timeStep}{Numeric giving the desired sampling frequency in
seconds. Default is
`NULL`, in which case `timeStep` corresponds to `0.8 / minPitch`.}

\item{start}{Numeric giving the desired start time of analysis in
seconds. Default is `NULL`, equivalent to the start of the audio file.}

\item{end}{Numeric giving the desired end time of analysis in seconds.
Default is `NULL`, equivalent to the duration of the audio file.}

\item{times}{Optional numeric vector giving the times to return signal
values from. Default is `NULL`.}

\item{minPitch}{Numeric giving the lowest pitch value in Hz to account for.
Default is `100`.}

\item{subtractMean}{Boolean; should mean pressure be subtracted around each
frame before applying Gaussian window? Default is `TRUE`.}

\item{toFile}{Boolean; if the output format is `ssff`, should these
objects be saved to disk as individual files? Default is `FALSE`.}

\item{outputDir}{String giving the location of the output directory of
SSFF files when `toFile = TRUE`. Default is to use the current working
directory.}

\item{outputExt}{String giving the file extension of SSFF files when
`toFile = TRUE`. Default is `.fmt`.}
}
\value{
A list of class `AsspDataObj` or data frame depending on the value
of `output`.
}
\description{
Wrapper for Praat procedure that estimates intensity (root-mean-squared
amplitude) values from an audio file or list of audio files.
}
\examples{
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
rms <- intensity(soundFile)
}
