% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity2df.R
\name{intensity2df}
\alias{intensity2df}
\title{Format Intensity object as data frame}
\usage{
intensity2df(rms, filename, times = NULL)
}
\arguments{
\item{rms}{`parselmouth.Intensity` object}

\item{filename}{String giving the file name of audio file.}

\item{times}{Optional numeric vector giving the times to return signal
values from.}
}
\value{
Data frame with root-mean-squared amplitude values.
}
\description{
Convert Praat Intensity object (as loaded in memory through the Python
interface `parselmouth`) to data frame.
}
\examples{
# Don't use directly
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
rms_df <- intensity(soundFile, output = 'df')
}
