% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensity2ssff.R
\name{intensity2ssff}
\alias{intensity2ssff}
\title{Format Intensity object as SSFF}
\usage{
intensity2ssff(rms)
}
\arguments{
\item{rms}{`parselmouth.Intensity` object}
}
\value{
List object of class `AsspDataObj`.
}
\description{
Convert Praat Intensity object (as loaded in memory through the Python
interface `parselmouth`) to `list` object with the Simple Signal File Format.
}
\examples{
# Don't use directly
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
rms_ssff <- intensity(soundFile, output = 'ssff')
}
