% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch2cand.R
\name{pitch2cand}
\alias{pitch2cand}
\title{Format Pitch object as long data frame with candidate information}
\usage{
pitch2cand(pit, filename)
}
\arguments{
\item{pit}{`parselmouth.Pitch` object}

\item{filename}{String giving the file name of audio file.}
}
\value{
Data frame with pitch candidate values.
}
\description{
Convert Praat Pitch object (as loaded in memory through the Python
interface `parselmouth`) to long-form data frame retaining information about
pitch candidates and their weights.
}
\examples{
# Don't use directly
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
pit_cand <- pitchRawAC(soundFile, output = 'candidates')
}
