% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch2df.R
\name{pitch2df}
\alias{pitch2df}
\title{Format Pitch object as data frame}
\usage{
pitch2df(pit, filename, times = NULL)
}
\arguments{
\item{pit}{`parselmouth.Pitch` object}

\item{filename}{String giving the file name of audio file.}

\item{times}{Optional numeric vector giving the times to return signal
values from.}
}
\value{
Data frame with pitch values.
}
\description{
Convert Praat Pitch object (as loaded in memory through the Python
interface `parselmouth`) to data frame.
}
\examples{
# Don't use directly
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
pit_df <- pitchRawAC(soundFile, output = 'df')
}
