% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitch2ssff.R
\name{pitch2ssff}
\alias{pitch2ssff}
\title{Format Pitch object as SSFF}
\usage{
pitch2ssff(pit)
}
\arguments{
\item{pit}{`parselmouth.Pitch` object}
}
\value{
List object of class `AsspDataObj`.
}
\description{
Convert Praat Pitch object (as loaded in memory through the Python
interface `parselmouth`) to `list` object with the Simple Signal File Format.
}
\examples{
# Don't use directly
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
pit_ssff <- pitchRawAC(soundFile, output = 'ssff')
}
