% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praatVersion.R
\name{praatVersion}
\alias{praatVersion}
\title{Get Praat version}
\usage{
praatVersion()
}
\value{
Used for side effects
}
\description{
Print the version of Praat used by the current `parselmouth` installation
}
\examples{
praatVersion()
}
