% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printInfo.R
\name{printInfo}
\alias{printInfo}
\title{Return file information}
\usage{
printInfo(file)
}
\arguments{
\item{file}{String giving the path to a file that can be read by Praat.}
}
\value{
Used for side effects
}
\description{
Prints the default Praat information about any file
}
\examples{
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
printInfo(soundFile)
}
