% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFormant.R
\name{readFormant}
\alias{readFormant}
\title{Read Formant file}
\usage{
readFormant(
  filename,
  output = "ssff",
  ext = ".Formant",
  times = NULL,
  returnBw = FALSE,
  returnIntensity = FALSE,
  toFile = FALSE,
  outputDir = getwd(),
  outputExt = ".fmt"
)
}
\arguments{
\item{filename}{String giving the path to a Formant file or a directory of
Formant files. Alternatively a vector of strings giving the paths to Formant
files.}

\item{output}{String giving the desired output format. Should be either
`ssff` (default) for a list with the Simple Signal File Format or `df` for a
data frame.}

\item{ext}{String giving the file extension for Formant files; default is
`.Formant`. Only used if `filename` is a directory.}

\item{times}{Optional numeric vector giving the times to return signal
values from. Default is `NULL`.}

\item{returnBw}{Boolean; should formant bandwidths be returned? Default is
`FALSE`.}

\item{returnIntensity}{Boolean; should intensity values be returned?
Default is `FALSE`.}

\item{toFile}{Boolean; if the output format is `ssff`, should these
objects be saved to disk as individual files? Default is `FALSE`.}

\item{outputDir}{String giving the location of the output directory of
SSFF files when `toFile = TRUE`. Default is to use the current working
directory.}

\item{outputExt}{String giving the file extension of SSFF files when
`toFile = TRUE`. Default is `.fmt`.}
}
\value{
A list of class `AsspDataObj` or data frame depending on the value
of `output`.
}
\description{
Read Formant file generated by Praat in SSFF or data frame format.
}
\examples{
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.Formant')
fmt <- readFormant(soundFile)
}
