% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHarmonicity.R
\name{readHarmonicity}
\alias{readHarmonicity}
\title{Read Harmonicity file}
\usage{
readHarmonicity(
  filename,
  output = "ssff",
  ext = ".Harmonicity",
  times = NULL,
  toFile = FALSE,
  outputDir = getwd(),
  outputExt = ".hnr"
)
}
\arguments{
\item{filename}{String giving the path to a Harmonicity file or a directory
of Harmonicity files. Alternatively a vector of strings giving the paths to
Harmonicity files.}

\item{output}{String giving the desired output format. Should be either
`ssff` (default) for a list with the Simple Signal File Format or `df` for a
data frame.}

\item{ext}{String giving the file extension for Harmonicity files; default is
`.Harmonicity`. Only used if `filename` is a directory.}

\item{times}{Optional numeric vector giving the times to return signal
values from. Default is `NULL`.}

\item{toFile}{Boolean; if the output format is `ssff`, should these
objects be saved to disk as individual files? Default is `FALSE`.}

\item{outputDir}{String giving the location of the output directory of
SSFF files when `toFile = TRUE`. Default is to use the current working
directory.}

\item{outputExt}{String giving the file extension of SSFF files when
`toFile = TRUE`. Default is `.hnr`.}
}
\value{
A list of class `AsspDataObj` or data frame depending on the value
of `output`.
}
\description{
Read Harmonicity file generated by Praat in SSFF or data frame format.
}
\examples{
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.Harmonicity')
hnr <- readHarmonicity(soundFile)
}
