% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIntensity.R
\name{readIntensity}
\alias{readIntensity}
\title{Read Intensity file}
\usage{
readIntensity(
  filename,
  output = "ssff",
  ext = ".Intensity",
  times = NULL,
  toFile = FALSE,
  outputDir = getwd(),
  outputExt = ".rms"
)
}
\arguments{
\item{filename}{String giving the path to a Intensity file or a directory
of Intensity files. Alternatively a vector of strings giving the paths to
Intensity files.}

\item{output}{String giving the desired output format. Should be either
`ssff` (default) for a list with the Simple Signal File Format or `df` for a
data frame.}

\item{ext}{String giving the file extension for Intensity files; default is
`.Intensity`. Only used if `filename` is a directory.}

\item{times}{Optional numeric vector giving the times to return signal
values from. Default is `NULL`.}

\item{toFile}{Boolean; if the output format is `ssff`, should these
objects be saved to disk as individual files? Default is `FALSE`.}

\item{outputDir}{String giving the location of the output directory of
SSFF files when `toFile = TRUE`. Default is to use the current working
directory.}

\item{outputExt}{String giving the file extension of SSFF files when
`toFile = TRUE`. Default is `.rms`.}
}
\value{
A list of class `AsspDataObj` or data frame depending on the value
of `output`.
}
\description{
Read Intensity file generated by Praat in SSFF or data frame format.
}
\examples{
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.Intensity')
rms <- readIntensity(soundFile)
}
