% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPitch.R
\name{readPitch}
\alias{readPitch}
\title{Read Pitch file}
\usage{
readPitch(
  filename,
  output = "ssff",
  ext = ".Pitch",
  times = NULL,
  toFile = FALSE,
  outputDir = getwd(),
  outputExt = ".f0"
)
}
\arguments{
\item{filename}{String giving the path to a Pitch file or a directory
of Pitch files. Alternatively a vector of strings giving the paths to
Pitch files.}

\item{output}{String giving the desired output format. Should be either
`ssff` (default) for a list with the Simple Signal File Format or `df` for a
data frame.}

\item{ext}{String giving the file extension for Pitch files; default is
`.Pitch`. Only used if `filename` is a directory.}

\item{times}{Optional numeric vector giving the times to return signal
values from. Default is `NULL`.}

\item{toFile}{Boolean; if the output format is `ssff`, should these
objects be saved to disk as individual files? Default is `FALSE`.}

\item{outputDir}{String giving the location of the output directory of
SSFF files when `toFile = TRUE`. Default is to use the current working
directory.}

\item{outputExt}{String giving the file extension of SSFF files when
`toFile = TRUE`. Default is `.f0`.}
}
\value{
A list of class `AsspDataObj` or data frame depending on the value
of `output`.
}
\description{
Read Pitch file generated by Praat in SSFF or data frame format.
}
\examples{
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.Pitch')
pit <- readPitch(soundFile)
}
