% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSound.R
\name{readSound}
\alias{readSound}
\title{Read audio file}
\usage{
readSound(filename, start = NULL, end = NULL)
}
\arguments{
\item{filename}{String giving the file name of audio file.}

\item{start}{Numeric giving the desired start time to load in
seconds. Default is `NULL`, equivalent to the start of the audio file.}

\item{end}{Numeric giving the desired end time to load in
Default is `NULL`, equivalent to the duration of the audio file.}
}
\value{
A list object
}
\description{
Read audio file with most extensions as a list object using Praat
}
\examples{
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
snd <- readSound(soundFile)
}
