% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTextGrid.R
\name{readTextGrid}
\alias{readTextGrid}
\title{Read TextGrid file}
\usage{
readTextGrid(filename, ext = ".TextGrid")
}
\arguments{
\item{filename}{String giving the path to a TextGrid file or a directory
of TextGrid files. Alternatively a vector of strings giving the paths to
TextGrid files.}

\item{ext}{String giving the file extension for TextGrid files; default is
`.TextGrid`. Only used if `filename` is a directory.}
}
\value{
A data frame
}
\description{
Read TextGrid file generated by Praat in data frame format.
}
\examples{
datapath <- system.file('extdata', package='praatutils')
tgFile <- paste0(datapath, '/1.TextGrid')
tg <- readTextGrid(tgFile)
}
