% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{spectrogram}
\alias{spectrogram}
\title{Generate spectrogram}
\usage{
spectrogram(
  filename,
  start = NULL,
  end = NULL,
  windowLength = 0.005,
  maxFreq = 5000,
  timeStep = 0.002,
  freqStep = 20,
  windowShape = "GAUSSIAN"
)
}
\arguments{
\item{filename}{String giving the path to an audio file.}

\item{start}{Numeric giving the desired start time of analysis in
seconds. Default is `NULL`, equivalent to the start of the audio file.}

\item{end}{Numeric giving the desired end time of analysis in seconds.
Default is `NULL`, equivalent to the duration of the audio file.}

\item{windowLength}{Numeric giving the duration of analysis windows in
seconds. Default is `0.005`.}

\item{maxFreq}{Numeric giving the maximum frequency subject to analysis.
Default is `5000`.}

\item{timeStep}{Numeric giving the desired sampling frequency in
seconds. Default is `0.002`.}

\item{freqStep}{Numeric giving the desired frequency resolution in Hz.
Default is `20`.}

\item{windowShape}{String giving the window shape to use for the analysis.
Default is `GAUSSIAN`; other options are `BARTLETT`, `HAMMING`, `HANNING`,
`SQUARE`, and `WELCH`.}
}
\value{
A matrix.
}
\description{
Wrapper for Praat procedure that generates a matrix of energy values by
frequency bin and time from an audio file using the fast Fourier transform.
}
\examples{
datapath <- system.file('extdata', package='praatutils')
soundFile <- paste0(datapath, '/1.wav')
spec <- spectrogram(soundFile)
}
